Cleaned up the matching code and hooked access level check.
We now do a basic check for the "Guests Management" module, as we
check if the access level is $WEB_ADMIN_ALL (defined in config.pm).

Refactored the matching code to accept a source id and an action id
to return. Also refactored the authentication code to return the source
id upon a successful authentication.